services/document/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
redis[hiredis]==5.0.1
aioboto3==12.3.0
aiofiles==23.2.1
python-multipart==0.0.6
grpcio==1.76.0
grpcio-tools==1.76.0
uuid7==0.1.0
